LayoutTests/js/regexp-norepeat-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=46077

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS re.test(str) is false
PASS successfullyParsed is true

TEST COMPLETE
